test(queue): split queue entry sorting tests to comply with LOC limit
Extract TestSortQueueEntries class from test_queue_ordering.py (423 lines)
to test_queue_entry_sorting.py (99 lines). This brings test_queue_ordering.py
down to 331 lines, under the 400-line CI threshold.

All 31 tests passing across both files.
